* tools/tiffsplit.c (tiffcp): TIFFGetField count field should be uint32 type for TIFFTAG_JPEGTABLES.  Patch by Christophe Deroulers.

+++ tools/tiffsplit.c
@@ -172,7 +172,7 @@ tiffcp(TIFF* in, TIFF* out)
 	CopyField(TIFFTAG_SAMPLESPERPIXEL, samplesperpixel);
 	CopyField(TIFFTAG_COMPRESSION, compression);
 	if (compression == COMPRESSION_JPEG) {
-		uint16 count = 0;
+		uint32 count = 0;
 		void *table = NULL;
 		if (TIFFGetField(in, TIFFTAG_JPEGTABLES, &count, &table)
 		    && count > 0 && table) {
